requirements.txt
faiss-cpu
pandas
tqdm
sentence-transformers
requests
streamlit